fixed the edit button
The edit button now is only displayed when the creator of the announcement is viewing the detail page. URL hack is still possible so that needs fixed still. also need to make sure that its ok to compare the announcement.creator with the request.user. They are comparing the model itself and I'm not sure that is best practice. will a change inside the model make the comparison fail?