Events and email updates
The Events app now can delete, rsvp, and send email and text messages to the rsvped users. 
I also changed the email settings in the settings.py file so that it uses an email account set up on spdmizzou.org. THis way, we can actually get emails instead of going to console.